requirements.txt
requests==2.28.1
fastapi==0.87.0
uvicorn==0.19.0
google-cloud-storage==2.6.0

# dev dependencies
pytest==7.1.3
httpx==0.23.0